refactor(backend): move ChatSessionHistoryEntry types to claude module
ChatSessionHistoryEntry and ChatSessionHistoryResponse are produced by
Claude transcript parsing logic, so they belong in the claude module
rather than projects. Consumers in the projects module now import
directly from claude/session-history.types.
