+++ backend/modules/claude/claude-session-history.service.ts
@@ -5,7 +5,7 @@ import HttpStatus from 'http-status';
 import { injectable } from 'inversify';
 import { filter, flatMap, map, pipe } from 'remeda';
 import { AppError } from '../../errors/AppError.js';
-import type { ChatSessionHistoryEntry } from '../projects/chat-session.types.js';
+import type { ChatSessionHistoryEntry } from './session-history.types.js';
 import { expandContentBlocks } from './content-block-expander.util.js';
 
 export interface ClaudeTranscriptHistory {
