+++ backend/modules/projects/chat-session.types.ts
@@ -30,19 +30,3 @@ export interface ListChatSessionsFilters {
   scopeType?: ChatSessionScopeType;
   scopeTargetId?: string;
 }
-
-export interface ChatSessionHistoryEntry {
-  role: string;
-  messageType: 'text' | 'tool' | 'tool_result' | 'user';
-  content: string;
-  toolName?: string;
-  toolId?: string;
-  isError?: boolean;
-  timestamp?: string;
-}
-
-export interface ChatSessionHistoryResponse {
-  sessionId: string;
-  claudeSessionId: string;
-  entries: ChatSessionHistoryEntry[];
-}
