+++ backend/modules/projects/chat-sessions.service.ts
@@ -7,8 +7,8 @@ import { ChatSessionHistoryResponseDto } from './dto/chat-session-history-respon
 import { ChatSessionSummaryDto } from './dto/chat-session-summary.dto.js';
 import { ChatSessionRepository } from './chat-session.repository.js';
 import { ProjectRepository } from './project.repository.js';
+import type { ChatSessionHistoryResponse } from '../claude/session-history.types.js';
 import type {
-  ChatSessionHistoryResponse,
   ChatSessionSummary,
   ChatSessionTargetType,
   ClaudeChatContext,
